+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+### 5.3.15
+- [Geometry] fixed GetPlane/GetTrafo for LinearRegression2d
 - [Math] Added Fun.IsPowerOfTwo, Fun.NextPowerOfTwo, Fun.PrevPowerOfTwo overloads for uint32 and uint64
 
 ### 5.3.14